INC-817: Moved `translate()` etc functions in their own file
These are used in different places inside the `service` package to convert
between `PrisonerIepLevel`/`IepDetail`, it makes sense for them to have
their own file and reuse them.
